Move `SubWindowCloseFilter` into the shared pilot GUI helpers
`_SubWindowCloseFilter` lived privately in
`solvcon/pilot/apps/obsrefl/_viewer.py`, where `DomainViewer` and
`LinePlotViewer` use it to learn synchronously when their
`QMdiSubWindow` closes, before Qt frees the widget it wraps.

The `Canvas` feature (`solvcon/pilot/canvas/`) needs the identical
pattern for its own cached 2D canvas widgets, but it is a core pilot
feature, not part of the `obsrefl` app, so it should not reach into
an app-specific module for it. Move the filter into
`solvcon/pilot/base/_gui_common.py`, the module both areas already
use for shared Qt helpers, and update `_viewer.py`'s two call sites
to `_gui_common.SubWindowCloseFilter`.
